+++ src/charts/PieChart.jsx
@@ -44,6 +44,14 @@ export default class PieChart extends ShallowComponent {
         let c = 2 * depth + 1;
         let mRadius = (((this.props.size - 1) / 2) / c) * ( c - depth - 1);
 
+        let root = depth > 1 ?
+            (<circle
+                fill="white"
+                cx={this.props.size / 2}
+                cy={this.props.size / 2}
+                r={mRadius}/>) :
+            undefined;
+
         return (
             <div className="rb-pie-chart" style={{width: this.props.size, height: this.props.size}}>
                 <div className="rb-pie-chart-back-tool" style={{display: this.state.clicked ? "inherit" : "none"}}>
@@ -51,11 +59,7 @@ export default class PieChart extends ShallowComponent {
                 </div>
                 <svg className="rb-pie-chart-svg">
                     {this.__renderPies(data, 360, 0, depth, depth - 1)}
-                    <circle
-                        fill="white"
-                        cx={this.props.size / 2}
-                        cy={this.props.size / 2}
-                        r={mRadius}/>
+                    {root}
                 </svg>
                 <div className="tooltip" id="tooltip"></div>
                 <Legend data={this.legends} width={this.props.size}/>